+++ crates/rome_js_analyze/tests/specs/nursery/useGroupedTypeImport/valid.ts
@@ -1,0 +1,28 @@
+import T from 'mod';
+
+import * as T from 'mod';
+
+import { T } from 'mod';
+
+import type { T } from 'mod';
+
+import type { T, U } from 'mod';
+
+import { type T, U } from 'mod';
+
+import { T, type U } from 'mod';
+
+import type T from 'mod';
+
+// unsupported syntax by Rome
+//import type T, { U } from 'mod';
+
+import T, { type U } from 'mod';
+
+import type * as T from 'mod';
+
+import 'mod';
+
+import {} from 'mod';
+
+import type { default as A } from "mod"