+++ CHANGELOG.md
@@ -13,12 +13,15 @@ output. [#4405](https://github.com/rome/tools/pull/4405)
 
 #### New rules
 
+- [`useGroupedTypeImport`](https://docs.rome.tools/lint/rules/useGroupedTypeImport/)
 - [`noUselessConstructor`](https://docs.rome.tools/lint/rules/noUselessConstructor/)
 - [`useLiteralEnumMembers`](https://docs.rome.tools/lint/rules/useLiteralEnumMembers/)
 
 #### Other changes
+
 - Add new command `rome migrate` the transform the configuration file `rome.json`
 when there are breaking changes.
+- Fix [#4348](https://github.com/rome/tools/issues/4348) that caused [`noNonNullAssertion`](https://docs.rome.tools/lint/rules/nononnullassertion/) to emit incorrect code action
 
 ### Configuration
 ### Editors